+++ Editor/EditorWindow_SceneView.cs
@@ -1,5 +1,4 @@
 ﻿using ImGuiNET;
-using OpenTK.Graphics.OpenGL;
 
 namespace Engine;
 
@@ -20,25 +19,18 @@ public class EditorWindow_SceneView : EditorWindow
 			ImGui.SetNextWindowSize(Camera.I.size + new Vector2(0, 50), ImGuiCond.Always);
 			ImGui.SetNextWindowPos(new Vector2(0, 0), ImGuiCond.Always, new Vector2(0, 0));
 			ImGui.Begin("Scene View",
-				ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoMove | ImGuiWindowFlags.NoResize |
-				ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse);
+			            ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoMove | ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse);
 
 			ImGui.SetCursorPosX(Camera.I.size.X / 2 - 150);
 
-			Vector4 activeColor = new Color(0.21f, 0.9f, 0.98f, 1f).ToVector4();
-			Vector4 inactiveColor = new Color(1f, 1f, 1f, 1f).ToVector4();
+			var activeColor = new Color(0.21f, 0.9f, 0.98f, 1f).ToVector4();
+			var inactiveColor = new Color(1f, 1f, 1f, 1f).ToVector4();
 			ImGui.PushStyleColor(ImGuiCol.Text, Physics.Running ? activeColor : inactiveColor);
-			bool physicsButtonClicked = ImGui.Button("physics");
+			var physicsButtonClicked = ImGui.Button("physics");
 			if (physicsButtonClicked)
 			{
-				if (Physics.Running == false)
-				{
-					Physics.StartPhysics();
-				}
-				else if (Physics.Running == true)
-				{
-					Physics.StopPhysics();
-				}
+				if (Physics.Running == false) Physics.StartPhysics();
+				else if (Physics.Running) Physics.StopPhysics();
 			}
 
 			ImGui.PopStyleColor();
@@ -47,25 +39,19 @@ public class EditorWindow_SceneView : EditorWindow
 
 			ImGui.PushStyleColor(ImGuiCol.Text, Global.GameRunning ? activeColor : inactiveColor);
 
-			bool playButtonClicked = ImGui.Button("play");
-			if (playButtonClicked)
-			{
-				Global.GameRunning = !Global.GameRunning;
-			}
+			var playButtonClicked = ImGui.Button("play");
+			if (playButtonClicked) Global.GameRunning = !Global.GameRunning;
 
 			ImGui.PopStyleColor();
 
 			ImGui.SameLine();
-			bool resetDataButtonClicked = ImGui.Button("delete data");
-			if (resetDataButtonClicked)
-			{
-				PersistentData.DeleteAll();
-			}
+			var resetDataButtonClicked = ImGui.Button("delete data");
+			if (resetDataButtonClicked) PersistentData.DeleteAll();
 
 			ImGui.SetCursorPosX(0);
 			Editor.sceneViewPosition = new Vector2(ImGui.GetCursorPosX(), ImGui.GetCursorPosY());
 			ImGui.Image((IntPtr) Window.I.postProcessRenderTexture.colorAttachment, Camera.I.size,
-				new Vector2(0, 1), new Vector2(1, 0));
+			            new Vector2(0, 1), new Vector2(1, 0));
 
 			ImGui.End();
 
@@ -77,14 +63,12 @@ public class EditorWindow_SceneView : EditorWindow
 			ImGui.SetNextWindowSize(Camera.I.size + new Vector2(0, 50), ImGuiCond.Always);
 			ImGui.SetNextWindowPos(new Vector2(0, 0), ImGuiCond.Always, new Vector2(0, 0));
 			ImGui.Begin("Scene View",
-				ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoMove | ImGuiWindowFlags.NoResize |
-				ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse |
-				ImGuiWindowFlags.NoDecoration);
+			            ImGuiWindowFlags.NoCollapse | ImGuiWindowFlags.NoMove | ImGuiWindowFlags.NoResize | ImGuiWindowFlags.NoScrollbar | ImGuiWindowFlags.NoScrollWithMouse | ImGuiWindowFlags.NoDecoration);
 
 			ImGui.SetCursorPosX(0);
 			Editor.sceneViewPosition = new Vector2(ImGui.GetCursorPosX(), ImGui.GetCursorPosY());
 			ImGui.Image((IntPtr) Window.I.postProcessRenderTexture.colorAttachment, Camera.I.size,
-				new Vector2(0, 1), new Vector2(1, 0));
+			            new Vector2(0, 1), new Vector2(1, 0));
 
 			ImGui.End();
 		}
